Move LineGraph into a separate file
